Warnings from date processing get passed into Response by DataPrepper
In DataPrepper, new method to call date processing and handle any
warnings. This also DRYs up repeated CSpaceDate.new code

In Dates, new `warnings?` method is added for use by DataPrepper and
warning Hashes are collected instead of warning Strings.

Apparently redundant change made to TermSearchable. Cache lookup was
set to search: true, which I thought was supposed to search the client
if the term was not found in cache. However that was not happening, at
least for tests. So now the client is explicitly searched if no cached
vocabulary term is found.