implement "the bomb"
If a field has a blank value, no node for that field exists in the XML document
returned by collectionspace-mapper. When this XML is uploaded to
update an existing record, only nodes present in the XML affect the
existin record. That is: uploading a CSV without a value for a given
field will NOT cause that field to be blank in existing CSpace
records.

If a field consists of the bomb emoji (https://emojipedia.org/bomb/)
then a blank-value node for that field will be in the XML document
produced by collectionspace-mapper. When **this** XML is sent to
CSpace to update a record, any existing value(s) in that field will be
replaced with a blank. That is: using the bomb will destroy any
existing content in a field.